3.8.1.66

+++ pilot/workflow/generic.py
@@ -19,55 +19,69 @@
 # Authors:
 # - Mario Lassnig, mario.lassnig@cern.ch, 2016-2017
 # - Daniel Drizhuk, d.drizhuk@gmail.com, 2017
-# - Paul Nilsson, paul.nilsson@cern.ch, 2017-23
+# - Paul Nilsson, paul.nilsson@cern.ch, 2017-24
 # - Shuwei Ye, yesw@bnl.gov, 2021
 
-from __future__ import print_function  # Python 2, 2to3 complains about this
-
 import functools
+import logging
 import signal
 import threading
 import traceback
 import queue
 
-from time import time, sleep
-from sys import stderr
+from collections import namedtuple
 from os import getpid
 from shutil import rmtree
-
-from collections import namedtuple
+from sys import stderr
+from time import (
+    time,
+    sleep
+)
+from types import FrameType
 
 from pilot.common.exception import ExcThread
-from pilot.control import job, payload, data, monitor
-from pilot.util.constants import SUCCESS, PILOT_KILL_SIGNAL, MAX_KILL_WAIT_TIME
-from pilot.util.processes import kill_processes, threads_aborted
+from pilot.util.constants import (
+    MAX_KILL_WAIT_TIME,
+    PILOT_KILL_SIGNAL,
+    SUCCESS,
+    FAILURE
+)
+from pilot.control import (
+    data,
+    job,
+    monitor,
+    payload,
+)
+from pilot.util.processes import (
+    kill_processes,
+    threads_aborted
+)
 from pilot.util.timing import add_to_pilot_timing
 
-import logging
 logger = logging.getLogger(__name__)
+# Define Traces namedtuple at the module level
+Traces = namedtuple("Traces", ["pilot"])
 
 
-def interrupt(args, signum, frame):
+def interrupt(args: object, signum: int, frame: FrameType):
     """
-    Interrupt function on the receiving end of kill signals.
+    Handle signals for graceful exit.
+
     This function is forwarded any incoming signals (SIGINT, SIGTERM, etc) and will set abort_job which instructs
     the threads to abort the job.
 
-    :param args: pilot arguments.
-    :param signum: signal.
-    :param frame: stack/execution frame pointing to the frame that was interrupted by the signal.
+    :param args: pilot arguments (object)
+    :param signum: signal number (int)
+    :param frame: stack/execution frame pointing to the frame that was interrupted by the signal (object).
     """
-
-    try:
-        sig = [v for v, k in signal.__dict__.iteritems() if k == signum][0]
-    except Exception:
-        sig = [v for v, k in list(signal.__dict__.items()) if k == signum][0]
+    sig = [v for v, k in list(signal.__dict__.items()) if k == signum][0]
 
     # ignore SIGUSR1 since that will be aimed at a child process
     #if str(sig) == 'SIGUSR1':
     #    logger.info('ignore intercepted SIGUSR1 aimed at child process')
     #    return
-
+    if not hasattr(args, 'signal_counter'):
+        args.signal_counter = 0
     args.signal_counter += 1
 
     # keep track of when first kill signal arrived, any stuck loops should abort at a defined cut off time
@@ -79,8 +93,9 @@ def interrupt(args, signum, frame):
     if args.kill_time and current_time - args.kill_time > max_kill_wait_time:
         logger.warning('passed maximum waiting time after first kill signal - will commit suicide - farewell')
         try:
-            rmtree(args.sourcedir)
-        except Exception as e:
+            if hasattr(args, 'sourcedir'):
+                rmtree(args.sourcedir)
+        except (TypeError, OSError) as e:
             logger.warning(e)
         logging.shutdown()
         kill_processes(getpid())
@@ -91,37 +106,46 @@ def interrupt(args, signum, frame):
 
     args.signal = sig
     logger.warning('will instruct threads to abort and update the server')
+
+    if not hasattr(args, 'abort_job'):
+        args.abort_job = threading.Event()
     args.abort_job.set()
+
     logger.warning('setting graceful stop (in case it was not set already)')
+
+    if not hasattr(args, 'graceful_stop'):
+        args.graceful_stop = threading.Event()
     args.graceful_stop.set()
+
     logger.warning('waiting for threads to finish')
+
+    if not hasattr(args, 'job_aborted'):
+        args.job_aborted = threading.Event()
     args.job_aborted.wait(timeout=180)
 
 
-def register_signals(signals, args):
+def register_signals(signals: list, args: object):
     """
     Register kill signals for intercept function.
 
-    :param signals: list of signals.
-    :param args: pilot args.
-    :return:
+    :param signals: list of signals (list)
+    :param args: pilot arguments object (object).
     """
-
     for sig in signals:
         signal.signal(sig, functools.partial(interrupt, args))
 
 
-def run(args):
+def run(args: object) -> Traces or None:
     """
     Main execution function for the generic workflow.
 
     The function sets up the internal queues which handle the flow of jobs.
 
-    :param args: pilot arguments.
-    :returns: traces.
+    :param args: pilot arguments object (object)
+    :returns: traces object (Traces namedtuple or None)
     """
-
     logger.info('setting up signal handling')
+
     register_signals([signal.SIGINT,
                       signal.SIGTERM,
                       signal.SIGQUIT,
@@ -166,15 +190,15 @@ def run(args):
     # queues.interceptor_messages = queue.Queue()
 
     logger.info('setting up tracing')
-    traces = namedtuple('traces', ['pilot'])
-    traces.pilot = {'state': SUCCESS,
-                    'nr_jobs': 0,
-                    'error_code': 0,
-                    'command': None}
+    # Initialize traces with default values
+    traces = Traces(pilot={"state": SUCCESS, "nr_jobs": 0, "error_code": 0, "command": None})
 
     # initial sanity check defined by pilot user
     try:
-        user = __import__('pilot.user.%s.common' % args.pilot_user.lower(), globals(), locals(),
+        if not hasattr(args, 'pilot_user'):
+            logger.warning('pilot_user not defined - setting generic user')
+            args.pilot_user = 'generic'
+        user = __import__(f'pilot.user.{args.pilot_user.lower()}.common', globals(), locals(),
                           [args.pilot_user.lower()], 0)
         exit_code = user.sanity_check()
     except Exception as exc:
@@ -182,10 +206,13 @@ def run(args):
     else:
         if exit_code != 0:
             logger.info('aborting workflow since sanity check failed')
-            traces.pilot['error_code'] = exit_code
+            # Update traces using _replace for immutable update
+            traces = traces._replace(pilot={"state": FAILURE,
+                                            "nr_jobs": traces.pilot["nr_jobs"],
+                                            "error_code": exit_code})
+            #traces.pilot['error_code'] = exit_code
             return traces
-        else:
-            logger.info('passed sanity check')
+        logger.info('passed sanity check')
 
     # define the threads
     targets = {'job': job.control, 'payload': payload.control, 'data': data.control, 'monitor': monitor.control}
@@ -193,15 +220,14 @@ def run(args):
                          name=name) for name, target in list(targets.items())]
 
     logger.info('starting threads')
-    [thread.start() for thread in threads]
+    _ = [thread.start() for thread in threads]
 
     logger.info('waiting for interrupts')
 
-    # the thread_count is the total number of threads, not just the ExcThreads above
-    thread_count = threading.activeCount()
+    # the active_count() is the total number of threads, not just the ExcThreads above
     abort = False
     try:
-        while threading.activeCount() > 1 or not abort:
+        while threading.active_count() > 1 or not abort:
             # Note: this loop only includes at ExcThreads, not MainThread or Thread
             # threading.activeCount() will also include MainThread and any daemon threads (will be ignored)
             for thread in threads:
@@ -211,7 +237,7 @@ def run(args):
                 except queue.Empty:
                     pass
                 else:
-                    exc_type, exc_obj, exc_trace = exc
+                    _, exc_obj, _ = exc
                     # deal with the exception
                     print(f'received exception from bucket queue in generic workflow: {exc_obj}', file=stderr)
 
@@ -221,9 +247,13 @@ def run(args):
             abort = threads_aborted(caller='run')
             if abort:
                 logger.debug('will proceed to set job_aborted')
+
+                if not hasattr(args, 'job_aborted'):
+                    args.job_aborted = threading.Event()
                 args.job_aborted.set()
+
                 sleep(5)  # allow monitor thread to finish (should pick up job_aborted within 1 second)
-                logger.debug(f'all relevant threads have aborted (thread count={threading.activeCount()})')
+                logger.debug(f'all relevant threads have aborted (thread count={threading.active_count()})')
                 break
 
             sleep(1)
